Post WM_DESTROY instead of calling DestroyWindow()
FlutterWindow parent class has some business logic to perform on close.
Calling DistroyWindow is part of that.
Doing as we previously did may be skipping part of that logic.
That could be a cause for the access violation we noticed in the dumps.
PostMessage won't wait the message to be handled,
thus the framework has the chance to do whatever it needs to.